+++ src/components/dialog-temp.tsx
@@ -1,22 +1,27 @@
 'use client';
 
-import { ReactNode } from 'react';
+import { ReactNode, useMemo } from 'react';
 import { Modal, ModalProps } from './modal-temp';
+import { ToggleProvider } from '@/providers/toggle-provider';
+import React from 'react';
 
 export function Dialog({
   children,
   title,
-  cancelButton,
-  confirmButton,
   ...props
 }: Omit<ModalProps, 'show'> & { confirmButton?: ReactNode; cancelButton?: ReactNode }) {
+  const childArray = useMemo(() => React.Children.toArray(children), [children]);
+  const confirmButton = childArray.find((c: any) => c.type === Dialog.ConfirmButton);
+  const cancelButton = childArray.find((c: any) => c.type === Dialog.CancelButton);
+  const content = childArray.find((c: any) => c.type === Dialog.Content);
+
   return (
     <Modal
       {...props}
       title={title}
       show={true}>
       <div className='flex flex-col w-full gap-4 p-4'>
-        {children}
+        {content}
         <div className='flex gap-2 w-full'>
           {cancelButton}
           {confirmButton}
@@ -25,3 +30,18 @@ export function Dialog({
     </Modal>
   );
 }
+
+Dialog.ConfirmButton = function ({
+  children,
+  action,
+}: React.PropsWithChildren & { action?: () => Promise<void> }) {
+  return <ToggleProvider.Trigger action={action}>{children}</ToggleProvider.Trigger>;
+};
+
+Dialog.CancelButton = function ({ children }: React.PropsWithChildren) {
+  return <ToggleProvider.Trigger>{children}</ToggleProvider.Trigger>;
+};
+
+Dialog.Content = function ({ children }: React.PropsWithChildren) {
+  return children;
+};