+++ src/features/geolocation/providers/geolocation-provider.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { getDistanceInMeters } from '@/features/distance/util/get-distance-in-meters';
 import { useSessionStorage } from '@/hooks/use-session-storage';
 import { createContextWithUseHook } from '@/util/create-context-with-use-hook';
 import { useEffect, useRef, useState } from 'react';
@@ -7,11 +8,18 @@ import toast from 'react-hot-toast';
 
 const [GeolocationContext, useGeolocationContext] = createContextWithUseHook<{
   position: GeolocationPosition | null;
+  distanceToPreviousPosition: number;
 }>('useGeolocationContext can only be used within the scope of a GeolocationContext!');
 
 /**Retrieves the current position of the device using the geolocation-api, and serves it through its context. */
 export function GeolocationProvider({ children }: React.PropsWithChildren) {
   const [position, setPosition] = useSessionStorage<GeolocationPosition | null>('incant-pos', null);
+  const [previousPosition, setPreviousPosition] = useState(null);
+
+  const distanceToPreviousPosition =
+    position && previousPosition
+      ? getDistanceInMeters(position.coords, previousPosition.coords)
+      : Infinity;
 
   useEffect(() => {
     if (typeof window === 'undefined') {
@@ -22,12 +30,16 @@ export function GeolocationProvider({ children }: React.PropsWithChildren) {
     if ('geolocation' in navigator) {
       const options: PositionOptions = {
         enableHighAccuracy: true,
-        maximumAge: 0,
-        timeout: Infinity,
+        //maximumAge: 0,
+        //timeout: Infinity,
       };
 
       geoWatcher = navigator.geolocation.watchPosition(
         pos => {
+          if (position) {
+            setPreviousPosition(position);
+          }
+
           setPosition(pos);
         },
         error => {
@@ -45,7 +57,11 @@ export function GeolocationProvider({ children }: React.PropsWithChildren) {
     };
   }, []);
 
-  return <GeolocationContext.Provider value={{ position }}>{children}</GeolocationContext.Provider>;
+  return (
+    <GeolocationContext.Provider value={{ position, distanceToPreviousPosition }}>
+      {children}
+    </GeolocationContext.Provider>
+  );
 }
 
 export { useGeolocationContext };